[Refactor] Restructure calculator project into modern MVC layered architecture
