+++ src/main/java/calculator/Application.java
@@ -1,9 +1,9 @@
 package calculator;
 
-import calculator.runner.CalculatorRunner;
+import calculator.controller.CalculatorController;
 
 public class Application {
     public static void main(String[] args) {
-        new CalculatorRunner().run();
+        new CalculatorController().run();
     }
 }
