+++ src/main/java/calculator/view/InputView.java
@@ -1,4 +1,4 @@
-package calculator.io;
+package calculator.view;
 
 import camp.nextstep.edu.missionutils.Console;
