+++ src/main/java/calculator/view/OutputView.java
@@ -1,4 +1,4 @@
-package calculator.io;
+package calculator.view;
 
 public class OutputView {
     private static final String RESULT_FORMAT = "결과 : %d";
